requirements.txt
Django==1.8.3
pymysql
djangorestframework
django-rest-swagger
